deb+snap packaging related

+++ apt_install_dart.sh
@@ -1,8 +1,16 @@
 :
 
-sudo apt install -y apt-transport-https                                                                              
-curl -o- https://dl-ssl.google.com/linux/linux_signing_key.pub | sudo gpg --dearmor -o /usr/share/keyrings/dart.gpg
-echo 'deb [signed-by=/usr/share/keyrings/dart.gpg] https://storage.googleapis.com/download.dartlang.org/linux/debian stable main' | sudo tee /etc/apt/sources.list.d/dart_stable.list
-sudo apt update             
-sudo apt install -y dart                
+key='/usr/share/keyrings/googlelinux.gpg'
+key_url='https://dl-ssl.google.com/linux/linux_signing_key.pub'
+repo_url='https://storage.googleapis.com/download.dartlang.org/linux/debian'
+list='/etc/apt/sources.list.d/dart_stable.list'
+arch="$(dpkg --print-architecture)"
+
+if [ ! -f "$key" ]; then
+	sudo apt install -y apt-transport-https
+	curl -o- "$key_url" | sudo gpg --dearmor -o "$key"
+	echo "deb [signed-by=$key arch=$arch] $repo_url stable main" | sudo tee "$list"
+	sudo apt update
+fi
+sudo apt install -y dart
